+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # avromatic changelog
 
 ## v0.21.0
-- Add support for additional parameters to
-  `AvroTurf::ConfluentSchemaRegistry#compatible?`.
+- Remove monkey-patches for `AvroTurf::ConfluentSchemaRegistry` and
+  `FakeConfluentSchemaRegistryServer` and depend on `avro_schema_registry-client`
+  instead.
 
 ## v0.20.0
 - Support schema stores with a `#clear_schemas` method.